+++ src/main/java/bob/commands/Command.java
@@ -1,3 +1,9 @@
+package bob.commands;
+
+import bob.storage.Storage;
+import bob.data.TaskList;
+import bob.ui.Ui;
+
 public abstract class Command {
 
     public abstract void execute(TaskList task, Ui ui, Storage storage);
