+++ src/main/java/bob/commands/Mark.java
@@ -1,3 +1,9 @@
+package bob.commands;
+
+import bob.storage.Storage;
+import bob.data.TaskList;
+import bob.ui.Ui;
+
 public class Mark extends Command {
     private final int index;
