+++ src/main/java/bob/data/BobException.java
@@ -1,3 +1,5 @@
+package bob.data;
+
 public class BobException extends Exception {
 
     public BobException(String message) {
