+++ src/main/java/bob/tasks/EventTask.java
@@ -1,3 +1,5 @@
+package bob.tasks;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
