+++ src/main/java/bob/tasks/Task.java
@@ -1,8 +1,10 @@
+package bob.tasks;
+
 import java.time.LocalDate;
 
 public abstract class Task {
     protected String description;
-    protected boolean isDone;
+    public boolean isDone;
 
     public Task(String description) {
         this.description = description;
