+++ text-ui-test/EXPECTED.TXT
@@ -25,4 +25,4 @@ Here are the tasks in your list:
 1.[T][ ] read book
 2.[D][ ] submit assignment (by: Sunday)
 3.[E][ ] team meeting (from: Mon 2pm to: 4pm)
-Bye. Hope to see you again soon!
+Commands.Bye. Hope to see you again soon!
